add CNAME file to root
there is a big with Github Pages in which case the Custom Domain option in the Github Pages settings gets overwritten each time a deployment is made. This is a workaround to that. In the future when this bug is fixed, we should no longer need this file. At the same time however, it is not causing any harm so we can just leave it indefinitely.